-Bug fix:Objects with equal address in CReg                                     (Serializer.h,Serializer.cpp)
-set<class *> replaced by list<class *> in serializable classes
 (CObject.h,CObject.cpp,FeatureHandler.h,FeatureHandler.cpp,UnitHandler.h,UnitHandler.cpp,
 UnitTracker.cpp,PlasmaRepulser.h,PlasmaRepulser.cpp,LuaSyncedCtrl.cpp,LuaSyncedRead.cpp,
 LuaUnsyncedRead.cpp,SpawnScript.cpp,Game.h,Game.cpp,SelectedUnits.h,SelectedUnits.cpp,Team.h,Team.cpp
 WaitCommandsAI.h,WaitCommandsAI.cpp,MiniMap.cpp,MouseHandler.cpp,SelectionKeyHandler.cpp)

-Added/fixed CReg metadata registration
 (GlobalStuff.h,GlobalStuff.cpp,Feature.h,Feature.cpp,FeatureHandler.h,FeatureHandler.cpp,
 TAAirMoveType.h,TAAirMoveType.cpp,AirMoveType.cpp,groundmovetype.cpp,ScriptMoveType.cpp,
 BeamLaserProjectile.cpp,EmgProjectile.cpp,ExplosionGenerator.h,ExplosionGenerator.cpp,
 FireBallProjectile.cpp,FlameProjectile.cpp,FlareProjectile.h,FlareProjectile.cpp,
 LargeBeamLaserProjectile.cpp,LaserProjectile.cpp,MissileProjectile.cpp,MuzzleFlame.h,
 MuzzleFlame.cpp,PieceProjectile.h,PieceProjectile.cpp,ProjectileHandler.h,ProjectileHandler.cpp,
 RepulseGfx.cpp,ShieldPartProjectile.cpp,StarburstProjectile.cpp,WeaponProjectile.h,WeaponProjectile.cpp,
 WreckProjectile.h,WreckProjectile.cpp,Unit.h,Unit.cpp,AirCAI.h,AirCAI.cpp,BuilderCAI.h,BuilderCAI.cpp,
 CommandAI.h,CommandAI.cpp,CommandQueue.h,FactoryCAI.h,FactoryCAI.cpp,MobileCAI.h,MobileCAI.cpp,
 TransportCAI.h,TransportCAI.cpp,Builder.h,Builder.cpp,Building.h,Building.cpp,ExtractorBuilding.h,
 ExtractorBuilding.cpp,Factory.h,Factory.cpp,TransportUnit.h,TransportUnit.cpp,bombdropper.cpp,
 TorpedoLauncher.cpp,Weapon.cpp,GroundFlash.h,GroundFlash.cpp,MetalMap.h,MetalMap.cpp,command.h,
 command.cpp,Game.h,Game.cpp,Player.h,Player.cpp,Team.h,Team.cpp,aikey.h,GlobalAI.h,GlobalAI.cpp,
 GlobalAIHandler.h,GlobalAIHandler.cpp,Group.h,Group.cpp,GroupHandler.h,GroupHandler.cpp)

-Fixed bug with death dependances from nonserialized CObject-based classes      (CObject.cpp)
-Reduced save size                                                              (LosHandler.h,LosHandler.cpp)

-Replaced type* by vector<type>
 (LosHandler.h,LosHandler.cpp,RadarHandler.h,RadarHandler.cpp,QuadField.h,QuadField.cpp,UnitHandler.h,
 UnitHandler.cpp,MetalMap.h,MetalMap.cpp,ReadMap.h,ReadMap.cpp,BaseGroundDrawer.cpp,GuiHandler.cpp,
 AICallback.h,AICallback.cpp,)

-Load/Save fixed                                                                (LoadScript.h,LoadScript.cpp)
-Directory for saves                                                            (Game.cpp,LoadScript.cpp)
-F8 now save to QuickSave.ssf                                                   (Game.cpp)
-Added game command .save [-y ]<savename>                                       (Game.cpp,WordCompletion.cpp)

-Added Save/Load interface for AI(ai versions incremented)
 (GlobalAI.h,GlobalAI.cpp,GlobalAIHandler.h,GlobalAIHandler.cpp,GroupHandler.h,GroupHandler.cpp,IGlobalAI.h,IGroupAI.h)

-Added metal/energy storage to CUnit                                            (Unit.h,Unit.cpp,CommanderScript.cpp)
-Fixed possible bug with time profiler                                          (UnitLoader.cpp)
-Ground flashes can be permanent                                                (GroundFlash.cpp)
-Log message text instead Loading step... (for save/load screen)                (myGL.cpp)
-Added CScript::ScriptSelected                                                  (Script.h,Script.cpp,LoadScript.h,LoadScript.cpp)
-Edited camera                                                                  (CameraController.cpp)
-Main camera draw moved from CGame::Draw to CGame::DrawWorld                    (Game.h,Game.cpp)

-Added group handler per team
 (Game.cpp,SelectedUnits.cpp,LuaUI.cpp,GlobalAICallback.cpp,AICallback.cpp,GroupHandler.cpp,GlobalAI.h,GlobalAI.cpp)

-Added max time diff                                                            (Game.cpp,GameServer.cpp)
-Added time profiling for draw                                                  (Game.cpp)
-Fixed load from CPreGame(doesn't work)                                         (PreGame.cpp)
-In end game box hide second stat by default                                    (EndGameBox.cpp)
-Draw minimap before interface                                                  (Game.cpp)
-Don't draw main camera if minimap is maximized                                 (Game.cpp,MiniMap.h,MiniMap.cpp)
-Added GetStartPos to CAICallback                                               (AICallback.h,AICallback.cpp,IAICallback.h)
-Fixed CCommandQueue in ai interface                                            (AICallback.h,IAICallback.h)
-Hide gaia team                                                                 (QuitBox.cpp,ShareBox.cpp)
-Add - between unit name and description                                        (SelectionKeyHandler.cpp)
-Added windows SEH support                                                      (main.cpp,seh.cpp,seh.h)
-Use crashrpt instead crashreport                                               (main.cpp)
-Delete #include<math.h>                                                        (FontTexture.cpp)
-Removed CUnit::loaded from c# interface                                        (IUnitDef.cs,IUnitDef_generated.h,IUnitDef_generated.cpp)

Notes by Tobi:

-Commented out: Added Land at 80 for air units (doesn't work)                                  (AirCAI.cpp)
-Use crashrpt instead of crashreport only if compiling on MSVC
-Fixed some extra qualifier GCC warnings in the patch
-Fixed some std::exceptions ctors being passed string literals, while std::exception doesn't support that with GCC
-Fixed a std::string being passed as argument to %s in a format string
-(streflop) Added SMath.h in libm/headers so it actually compiles (wonder why it compiled once but not second time)
-Fixed some float to int cast warnings (in flozi's latest commit)
-Removed some unused variables (not necessarily in patched code)

-TODO: Excluded rts/build/vstudio8 as I cant test that now (on linux)
-TODO: Circular radar needs checking, I'm not sure it still works
-TODO: The change from std::set to std::list needs to be looked at and probably refactored as it may have performance impact
 A 'unit_set' encapsulating std::map<int,CUnit*> would be a good solution probably (with int = unit ID)




git-svn-id: https://spring.clan-sy.com/svn/spring/trunk@3826 37977431-3df6-0310-b722-df95706aa16b

+++ rts/Sim/Units/COB/CobInstance.cpp
@@ -1082,11 +1082,11 @@ int CCobInstance::GetUnitVal(int val, int p1, int p2, int p3, int p4)
 	case SEISMIC_RADIUS:
 		return unit->seismicRadius;
 	case DO_SEISMIC_PING:
-		if (p1 == 0) p1 = unit->seismicSignature;
+		if (p1 == 0) p1 = (int)unit->seismicSignature;
 		unit->DoSeismicPing(p1);
 		break;
 	case CURRENT_FUEL:
-		return unit->currentFuel * COBSCALE;
+		return (int)(unit->currentFuel * COBSCALE);
 	case TRANSPORT_ID:
 		return unit->transporter?unit->transporter->id:-1;
 	default:
